Added check if condaenv exists

+++ R/zzz.R
@@ -1,19 +1,21 @@
 .onLoad <- function(libname, pkgname){
-             reticulate::conda_create("cellanneal-r", "python==3.8")
-             reticulate::conda_install("cellanneal-r", c("numpy==1.24", "scipy==1.9",
-                                                         "matplotlib==3.7", "pandas==1.5",
-                                                         "seaborn==0.12"))
-             reticulate::conda_install("cellanneal-r",
-                                       c("openpyxl==3.0",
-                                         "xlrd==1.2"),
-                                       channel = "anaconda")
-             reticulate::use_condaenv("cellanneal-r")
-             curl::curl_download("https://github.com/LiBuchauer/cellanneal/archive/refs/heads/master.zip", destfile = "./cellanneal.zip")
-             unzip("cellanneal.zip")
-             reticulate::py_install("cellanneal-master/", pip = TRUE)
-             file.remove("cellanneal.zip")
-             unlink("cellanneal-master", recursive = TRUE)
-             reticulate::use_condaenv("cellanneal-r")
+              if(!reticulate::condaenv_exists(cellanneal-r)){
+                reticulate::conda_create("cellanneal-r", "python==3.8")
+                reticulate::conda_install("cellanneal-r", c("numpy==1.24", "scipy==1.9",
+                                                             "matplotlib==3.7", "pandas==1.5",
+                                                             "seaborn==0.12"))
+                reticulate::conda_install("cellanneal-r",
+                                           c("openpyxl==3.0",
+                                             "xlrd==1.2"),
+                                           channel = "anaconda")
+                reticulate::use_condaenv("cellanneal-r")
+                curl::curl_download("https://github.com/LiBuchauer/cellanneal/archive/refs/heads/master.zip", destfile = "./cellanneal.zip")
+                unzip("cellanneal.zip")
+                reticulate::py_install("cellanneal-master/", pip = TRUE)
+                file.remove("cellanneal.zip")
+                unlink("cellanneal-master", recursive = TRUE)
+              }
+            reticulate::use_condaenv("cellanneal-r")
             reticulate::source_python(
               system.file("python/run_cellanneal.py",
                           package = "deconfectHelpers"),
